api/requirements.txt
hurry.filesize==0.9

fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6

Pillow==10.1.0